# docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.ChangePasswordRequest:
    properties:
      new_password:
        minLength: 8
        type: string
      old_password:
        type: string
    required:
    - new_password
    - old_password
    type: object
...
  handlers.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  handlers.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  response.Meta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  response.Pagination:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
  response.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: API cho nền tảng đọc truyện web novel Nekozanedex
  title: Nekozanedex API
  version: "1.0"
paths:
...
  /api/auth/change-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Password Info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Đổi mật khẩu
      tags:
      - Auth
  /api/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login Info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Đăng nhập
      tags:
      - Auth
  /api/auth/logout:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Đăng xuất
      tags:
      - Auth
  /api/auth/logout-all:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Đăng xuất tất cả thiết bị
      tags:
      - Auth
  /api/auth/profile:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Lấy thông tin profile
      tags:
      - Auth
...
  /api/auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Register Info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
      summary: Đăng ký tài khoản mới
      tags:
      - Auth
  /api/auth/sessions:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Lấy danh sách sessions đang active
      tags:
      - Auth
  /api/bookmarks:
    get:
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Pagination'
      security:
      - BearerAuth: []
      summary: Lấy danh sách bookmark của tôi
      tags:
      - Bookmarks
  /api/bookmarks/{storyId}:
    delete:
      parameters:
      - description: Story ID
        in: path
        name: storyId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Xóa bookmark
      tags:
      - Bookmarks
    post:
      parameters:
      - description: Story ID
        in: path
        name: storyId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Thêm bookmark
      tags:
      - Bookmarks
  /api/bookmarks/{storyId}/check:
    get:
      parameters:
      - description: Story ID
        in: path
        name: storyId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Kiểm tra đã bookmark chưa
      tags:
      - Bookmarks
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer token authentication. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"